+++ src/components/Icons/iconChangelogData.ts
@@ -15,6 +15,7 @@ export const changelogData: ChangelogData[] = [
     type: "Update",
     affects: ["Functionality"],
     notes: [
+      "Updated `iconColorsArray` to include `ui.warning.tertiary`.",
       "Added the `actionTaskAlt`, `communicationDynamicFeed`, `contentTypeAudio`, `contentTypeBook`, `contentTypeEBook`, `contentTypeEvent`, `contentTypeImage`, `contentTypeMisc`, `contentTypeVideo`, and `hardwareLaptop` icons.",
     ],
   },